Split Backend.js into app.js, routes, and server.js
